requirements.txt
Numpy==1.19.2
IPython==7.19.0
Pillow==8.0.1
Torch==0.4.1
Torchvision==0.2.2
Matplotlib==3.3.2